pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rond-api"
version = "0.1.0"
description = "An API tool to connect with Apple app Rond(macOS Container)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Circle Crop" }]
dependencies = ["sqlalchemy>=2.0", "aiosqlite>=0.20", "python-dotenv>=1.0"]

[project.scripts]
rond-api = "rond_api.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
